#10 - 합성 애노테이션 적용
테스트 코드에서 합성 애노테이션을 적용하는 방법을 알았으니, 이제는 필요없어진 코드를 지워주었음

합성 애노테이션을 활용해서 초기 Spring Initializr 에서 생성해주었던 초기 세팅과 유사하게 설정해줄 수 있었음
