+++ src/test/java/com/tobyspring/helloboot/HelloServiceTest.java
@@ -8,23 +8,9 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.METHOD)
-@UnitTest // UnitTest 는 메타 애노테이션으로 적용할 수 없음 -> @Target 을 추가해야함
-@interface FastUnitTest {
-
-}
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.METHOD, ElementType.ANNOTATION_TYPE})
-@Test // meta - annotation
-@interface UnitTest {
-
-}
-
 public class HelloServiceTest {
 
-    @UnitTest
+    @Test
     void simpleHelloTestService() {
         SimpleHelloService simpleHelloService = new SimpleHelloService();
